+++ lib/ui/pages/main_page.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import './tab_page.dart';
 
 class MainPage extends StatefulWidget {
   @override
@@ -31,28 +32,6 @@ class _MainPageState extends State<MainPage> {
               TabPage(title: "ダッシュボード", icon: Icons.dashboard),
             ]
         ),
-      ),
-    );
-  }
-}
-
-class TabPage extends StatelessWidget {
-
-  final IconData icon;
-  final String title;
-
-  const TabPage({key, required this.icon, required this.title}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    final TextStyle? textStyle = Theme.of(context).textTheme.headline4;  // 文字のスタイル
-    return Center(
-      child:Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: <Widget>[
-          Icon(icon, size: 72.0, color: textStyle!.color),
-          Text(title, style: textStyle),
-        ],
       ),
     );
   }
